% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-transforms.R
\name{col2hcl}
\alias{col2hcl}
\title{Convert colour to HCL values}
\usage{
col2hcl(colour)
}
\arguments{
\item{colour}{a colour name from \code{\link[=colours]{colours()}} or \code{\link[=colors]{colors()}}}
}
\value{
A matrix of hue, chroma, and luminance
}
\description{
go from a colour like "red" to a HCL matrix, values of hue,
chroma, and luminance
}
\examples{
col2hcl("red")
}
