% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-transforms.R
\name{col2hex}
\alias{col2hex}
\title{Convert colour to hex value}
\usage{
col2hex(colour)
}
\arguments{
\item{colour}{a colour name from \code{\link[=colours]{colours()}} or \code{\link[=colors]{colors()}}}
}
\value{
A hex color hex value
}
\description{
go from a colour like "red" to the hex code, "#FF0000"
}
\examples{
col2hex("red")
}
