% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-transforms.R
\name{col2rgb}
\alias{col2rgb}
\title{Convert colour to RGB values}
\usage{
col2rgb(colour)
}
\arguments{
\item{colour}{a colour name from \code{\link[=colours]{colours()}} or \code{\link[=colors]{colors()}}}
}
\value{
A matrix of red, green, and blue.
}
\description{
go from a colour like "red" to a HCL matrix, values of red,
green, and blue.
}
\examples{
col2rgb("red")
}
