% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-transforms.R
\name{colour_data}
\alias{colour_data}
\title{Create a data frame of colour values}
\usage{
colour_data(colour)
}
\arguments{
\item{colour}{a colour name from \code{\link[=colours]{colours()}} or \code{\link[=colors]{colors()}}}
}
\value{
data frame of colours with columns: name, hex, hue, chroma,
luminance, red, green, blue
}
\description{
Create a data frame of colour values
}
\examples{
colour_data("red")
}
