% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-colour.R
\name{nearest_colour}
\alias{nearest_colour}
\title{Find closest named colours.}
\usage{
nearest_colour(
  colour,
  n = 1,
  method = c("hcl", "hue_chroma", "hue_luminance", "chroma_luminance", "hue", "chroma",
    "luminance")
)
}
\arguments{
\item{colour}{A character string specifying a color in hex format (e.g., "#FF5733") or a named R color (e.g., "red").}

\item{n}{An integer specifying the number of closest colors to return. Default is 1.}

\item{method}{character how to perform the sorting. Default is "hcl", and it is done by euclidean distance. Total options are "hcl", "hue_chroma",
"hue_luminance", "chroma_luminance", "hue", "chroma", and "luminance".}
}
\value{
A tibble with the closest colours and its details.
}
\description{
Searches through R's built-in named colors from \code{colors()} to find the closest matches to a given input color. Uses Euclidean distance in HCL (Hue-Chroma-Luminance) color space to determine similarity.
}
\examples{
library(colorspace)
qual_cols <- qualitative_hcl(7)
qual_cols
swatchplot(qual_cols)
near_qual_cols <- nearest_colour(qual_cols)
near_qual_cols$hex.x
near_qual_cols$hex.x |> swatchplot(qual_cols)
near_qual_cols$hex.x |> specplot(qual_cols)
nearest_colour(qual_cols)$hex.x |> swatchplot(qual_cols)
nearest_colour(qual_cols, method = "hue_chroma")$hex.x |> swatchplot(qual_cols)
nearest_colour(qual_cols, method = "hue_luminance")$hex.x |> swatchplot(qual_cols)

}
